+++ Editor/MemorySnapshot/Cached/Managed/ReferenceArrayCrawlerJobChunk.cs
@@ -5,6 +5,11 @@ using Unity.Collections;
 using Unity.MemoryProfiler.Editor.Containers;
 using Unity.MemoryProfiler.Editor.Format;
 using static Unity.MemoryProfiler.Editor.CachedSnapshot;
+#if !UNMANAGED_NATIVE_HASHMAP_AVAILABLE
+using AddressToManagedIndexHashMap = Unity.MemoryProfiler.Editor.Containers.CollectionsCompatibility.NativeHashMap<ulong, long>;
+#else
+using AddressToManagedIndexHashMap = Unity.Collections.NativeHashMap<ulong, long>;
+#endif
 
 namespace Unity.MemoryProfiler.Editor.Managed
 {
@@ -29,11 +34,11 @@ namespace Unity.MemoryProfiler.Editor.Managed
             readonly ManagedMemorySectionEntriesCache m_ManagedHeapBytes;
             readonly VirtualMachineInformation m_VMInfo;
             [ReadOnly]
-            readonly NativeHashMap<ulong, long> m_MangedObjectIndexByAddress;
+            readonly AddressToManagedIndexHashMap m_MangedObjectIndexByAddress;
             int m_ArrayObjectTypeIndex;
 
             public ReferenceArrayCrawlerJobChunk(
-                in ManagedMemorySectionEntriesCache managedHeapBytes, in VirtualMachineInformation vmInfo, in NativeHashMap<ulong, long> mangedObjectIndexByAddress,
+                in ManagedMemorySectionEntriesCache managedHeapBytes, in VirtualMachineInformation vmInfo, in AddressToManagedIndexHashMap mangedObjectIndexByAddress,
                 SourceIndex arrayObjectIndex,
                 ref DynamicArray<StackCrawlData> resultingCrawlDataStack, BytesAndOffset arrayData,
                  uint pointerSize, long startArrayIndex, long chunkElementCount, int arrayObjectTypeIndex)